requirements.txt
mlx-vlm
torch
transformers
Pillow
soundfile
minicpmo-utils[all]
librosa
mss
sounddevice
fastapi
uvicorn
onnx
decord2
